requirements.txt
torch
torchvision
numpy
pandas
geopandas
rasterio
rioxarray
xarray
scipy
matplotlib
opencv-python
fastapi
uvicorn
streamlit
folium
streamlit-folium
pillow
pickle5